+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Compatibility:
 
 Features:
   - Fixes a 500 error when custom styles is empty
+  - Fixes removal of existing css boxes when other configuration are changed
 
 v0.6.3
 ------